fix: building on windows
